pyproject.toml
[project]
name = "futures-bot"
version = "0.1.0"
description = "Professional modular Binance Futures trading system for small accounts"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.2",
    "python-dotenv>=1.1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
